+++ pkg/queryfrontend/roundtrip_test.go
@@ -16,6 +16,7 @@ import (
 	cortexcache "github.com/cortexproject/cortex/pkg/chunk/cache"
 	"github.com/cortexproject/cortex/pkg/ingester/client"
 	"github.com/cortexproject/cortex/pkg/querier/queryrange"
+	cortexvalidation "github.com/cortexproject/cortex/pkg/util/validation"
 	"github.com/go-kit/kit/log"
 	"github.com/prometheus/prometheus/pkg/timestamp"
 	"github.com/prometheus/prometheus/promql/parser"
@@ -31,19 +32,10 @@ const (
 	day     = 24 * time.Hour
 )
 
-// fakeLimits implements the Cortex queryrange.Limits interface.
-type fakeLimits struct{}
-
-func (l *fakeLimits) MaxQueryLength(_ string) time.Duration {
-	return 7 * 24 * time.Hour
-}
-
-func (l *fakeLimits) MaxQueryParallelism(_ string) int {
-	return 14
-}
-
-func (l *fakeLimits) MaxCacheFreshness(_ string) time.Duration {
-	return time.Minute
+var defaultLimits = &cortexvalidation.Limits{
+	MaxQueryLength:      7 * 24 * time.Hour,
+	MaxQueryParallelism: 14,
+	MaxCacheFreshness:   time.Minute,
 }
 
 // fakeRoundTripper implements the RoundTripper interface.
@@ -83,8 +75,6 @@ func TestRoundTripRetryMiddleware(t *testing.T) {
 		End:   2 * hour,
 		Step:  10 * seconds,
 	}
-
-	codec := NewThanosCodec(true)
 
 	for _, tc := range []struct {
 		name       string
@@ -139,8 +129,13 @@ func TestRoundTripRetryMiddleware(t *testing.T) {
 	} {
 
 		t.Run(tc.name, func(t *testing.T) {
-			tpw, err := NewTripperWare(&fakeLimits{}, nil, codec, nil,
-				day, tc.maxRetries, nil, log.NewNopLogger())
+			tpw, err := NewTripperware(
+				Config{
+					SplitQueriesByInterval: day,
+					MaxRetries:             tc.maxRetries,
+					CortexLimits:           defaultLimits,
+				}, nil, log.NewNopLogger(),
+			)
 			testutil.Ok(t, err)
 
 			rt, err := newFakeRoundTripper()
@@ -149,7 +144,7 @@ func TestRoundTripRetryMiddleware(t *testing.T) {
 			rt.setHandler(handler)
 
 			ctx := user.InjectOrgID(context.Background(), "1")
-			httpReq, err := codec.EncodeRequest(ctx, tc.req)
+			httpReq, err := NewThanosCodec(true).EncodeRequest(ctx, tc.req)
 			testutil.Ok(t, err)
 
 			_, err = tpw(rt).RoundTrip(httpReq)
@@ -227,8 +222,12 @@ func TestRoundTripSplitIntervalMiddleware(t *testing.T) {
 	} {
 
 		t.Run(tc.name, func(t *testing.T) {
-			tpw, err := NewTripperWare(&fakeLimits{}, nil, codec, nil,
-				tc.splitInterval, 0, nil, log.NewNopLogger())
+			tpw, err := NewTripperware(
+				Config{
+					SplitQueriesByInterval: tc.splitInterval,
+					CortexLimits:           defaultLimits,
+				}, nil, log.NewNopLogger(),
+			)
 			testutil.Ok(t, err)
 
 			rt, err := newFakeRoundTripper()
@@ -309,11 +308,14 @@ func TestRoundTripCacheMiddleware(t *testing.T) {
 		},
 	}
 
-	codec := NewThanosCodec(true)
-
 	now := time.Now()
-	tpw, err := NewTripperWare(&fakeLimits{}, cacheConf, codec, queryrange.PrometheusResponseExtractor{},
-		day, 0, nil, log.NewNopLogger())
+	tpw, err := NewTripperware(
+		Config{
+			SplitQueriesByInterval:   day,
+			CortexResultsCacheConfig: cacheConf,
+			CortexLimits:             defaultLimits,
+		}, nil, log.NewNopLogger(),
+	)
 	testutil.Ok(t, err)
 
 	rt, err := newFakeRoundTripper()
@@ -360,7 +362,7 @@ func TestRoundTripCacheMiddleware(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 
 			ctx := user.InjectOrgID(context.Background(), "1")
-			httpReq, err := codec.EncodeRequest(ctx, tc.req)
+			httpReq, err := NewThanosCodec(true).EncodeRequest(ctx, tc.req)
 			testutil.Ok(t, err)
 
 			_, err = tpw(rt).RoundTrip(httpReq)